MAINT: refactor- change dataloader to wsi
