+++ tests/test_wsireader.py
@@ -39,7 +39,6 @@ def strictly_increasing(seq):
 
     Returns:
         bool: True if strictly increasing.
-
     """
     return all(a < b for a, b in zip(seq, seq[1:]))
 
@@ -53,7 +52,6 @@ def strictly_decreasing(seq):
 
     Returns:
         bool: True if strictly decreasing.
-
     """
     return all(a > b for a, b in zip(seq, seq[1:]))
 
@@ -369,7 +367,6 @@ def test_find_optimal_level_and_downsample_openslide_interpolation_warning(
 
     This tests the case where the scale is found to be > 1 and interpolation
     will be applied to the output. A UserWarning should be raised in this case.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     with pytest.warns(UserWarning):
@@ -381,7 +378,6 @@ def test_find_optimal_level_and_downsample_jp2_interpolation_warning(_sample_jp2
 
     This tests the case where the scale is found to be > 1 and interpolation
     will be applied to the output. A UserWarning should be raised in this case.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     with pytest.warns(UserWarning):
@@ -481,7 +477,6 @@ def test_read_rect_openslide_baseline(_sample_ndpi):
     """Test openslide read rect at baseline.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     location = NDPI_TEST_TISSUE_LOCATION
@@ -497,7 +492,6 @@ def test_read_rect_jp2_baseline(_sample_jp2):
     """Test jp2 read rect at baseline.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     location = JP2_TEST_TISSUE_LOCATION
@@ -513,7 +507,6 @@ def test_read_rect_openslide_levels(_sample_ndpi):
     """Test openslide read rect with resolution in levels.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     location = NDPI_TEST_TISSUE_LOCATION
@@ -530,7 +523,6 @@ def test_read_rect_jp2_levels(_sample_jp2):
     """Test jp2 read rect with resolution in levels.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     location = (0, 0)
@@ -568,7 +560,6 @@ def test_read_rect_openslide_mpp(_sample_ndpi):
     """Test openslide read rect with resolution in microns per pixel.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     location = NDPI_TEST_TISSUE_LOCATION
@@ -580,7 +571,6 @@ def test_read_rect_jp2_mpp(_sample_jp2):
     """Test jp2 read rect with resolution in microns per pixel.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     location = JP2_TEST_TISSUE_LOCATION
@@ -592,7 +582,6 @@ def test_read_rect_openslide_objective_power(_sample_ndpi):
     """Test openslide read rect with resolution in objective power.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     location = NDPI_TEST_TISSUE_LOCATION
@@ -605,7 +594,6 @@ def test_read_rect_jp2_objective_power(_sample_jp2):
     """Test jp2 read rect with resolution in objective power.
 
     Location coordinate is in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     location = JP2_TEST_TISSUE_LOCATION
@@ -618,7 +606,6 @@ def test_read_bounds_openslide_baseline(_sample_ndpi):
     """Test openslide read bounds at baseline.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     bounds = NDPI_TEST_TISSUE_BOUNDS
@@ -634,7 +621,6 @@ def test_read_bounds_jp2_baseline(_sample_jp2):
     """Test jp2 read bounds at baseline.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     bounds = JP2_TEST_TISSUE_BOUNDS
@@ -655,7 +641,6 @@ def test_read_bounds_openslide_levels(_sample_ndpi):
     """Test openslide read bounds with resolution in levels.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     bounds = NDPI_TEST_TISSUE_BOUNDS
@@ -675,7 +660,6 @@ def test_read_bounds_jp2_levels(_sample_jp2):
     """Test jp2 read bounds with resolution in levels.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     bounds = JP2_TEST_TISSUE_BOUNDS
@@ -696,7 +680,6 @@ def test_read_bounds_openslide_mpp(_sample_ndpi):
     """Test openslide read bounds with resolution in microns per pixel.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     bounds = NDPI_TEST_TISSUE_BOUNDS
@@ -709,7 +692,6 @@ def test_read_bounds_jp2_mpp(_sample_jp2):
     """Test jp2 read bounds with resolution in microns per pixel.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     bounds = JP2_TEST_TISSUE_BOUNDS
@@ -722,7 +704,6 @@ def test_read_bounds_openslide_objective_power(_sample_ndpi):
     """Test openslide read bounds with resolution in objective power.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_ndpi)
     bounds = NDPI_TEST_TISSUE_BOUNDS
@@ -736,7 +717,6 @@ def test_read_bounds_jp2_objective_power(_sample_jp2):
     """Test jp2 read bounds with resolution in objective power.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OmnyxJP2WSIReader(_sample_jp2)
     bounds = JP2_TEST_TISSUE_BOUNDS
@@ -750,7 +730,6 @@ def test_read_bounds_interpolated(_sample_svs):
     """Test openslide read bounds with interpolated output.
 
     Coordinates in baseline (level 0) reference frame.
-
     """
     wsi = wsireader.OpenSlideWSIReader(_sample_svs)
     bounds = SVS_TEST_TISSUE_BOUNDS
