+++ tiatoolbox/__init__.py
@@ -20,7 +20,7 @@
 
 """Top-level package for TIA Toolbox."""
 from tiatoolbox import tiatoolbox
-from tiatoolbox import dataloader
+from tiatoolbox import wsi
 from tiatoolbox import utils
 from tiatoolbox import tools
