+++ tiatoolbox/cli.py
@@ -20,7 +20,7 @@
 
 """Console script for tiatoolbox."""
 from tiatoolbox import __version__
-from tiatoolbox import dataloader
+from tiatoolbox import wsi
 from tiatoolbox.tools import stainnorm as sn
 from tiatoolbox import utils
 from tiatoolbox.utils.exceptions import MethodNotSupported
@@ -100,9 +100,7 @@ def slide_info(wsi_input, output_dir, file_types, mode, verbose=True):
         output_dir.mkdir(parents=True, exist_ok=True)
 
     for curr_file in files_all:
-        slide_param = dataloader.slide_info.slide_info(
-            input_path=curr_file, verbose=verbose
-        )
+        slide_param = wsi.slide_info.slide_info(input_path=curr_file, verbose=verbose)
         if mode == "show":
             print(slide_param.as_dict())
 
@@ -157,7 +155,7 @@ def read_bounds(wsi_input, region, resolution, units, output_path, mode):
         input_dir = pathlib.Path(wsi_input).parent
         output_path = str(input_dir.parent / "im_region.jpg")
 
-    wsi = dataloader.wsireader.get_wsireader(input_img=wsi_input)
+    wsi = wsi.wsireader.get_wsireader(input_img=wsi_input)
 
     im_region = wsi.read_bounds(
         region,
@@ -191,7 +189,7 @@ def slide_thumbnail(wsi_input, output_path, mode):
         input_dir = pathlib.Path(wsi_input).parent
         output_path = str(input_dir.parent / "slide_thumb.jpg")
 
-    wsi = dataloader.wsireader.get_wsireader(input_img=wsi_input)
+    wsi = wsi.wsireader.get_wsireader(input_img=wsi_input)
 
     slide_thumb = wsi.slide_thumbnail()
 
@@ -258,7 +256,7 @@ def save_tiles(
     print(files_all)
 
     for curr_file in files_all:
-        dataloader.save_tiles.save_tiles(
+        wsi.save_tiles.save_tiles(
             input_path=curr_file,
             output_dir=output_dir,
             tile_objective_value=tile_objective_value,
